backend/requirements.txt
flask==2.2.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.2.1
pandas==1.4.2
numpy==1.22.4
nltk==3.8.1
joblib==1.3.2
spacy==3.7.4
werkzeug==2.2.3 